feat(api): move TabCompleter and UserStockHolderOperator to api module
