+++ api/src/main/java/net/okocraft/box/api/util/TabCompleter.java
@@ -1,4 +1,4 @@
-package net.okocraft.box.feature.command.util;
+package net.okocraft.box.api.util;
 
 import net.okocraft.box.api.BoxProvider;
 import org.bukkit.Bukkit;
@@ -9,8 +9,17 @@ import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 
+/**
+ * A utility class for tab completions.
+ */
 public final class TabCompleter {
 
+    /**
+     * Gets the set of item names that match the filter.
+     *
+     * @param filter the item name filter
+     * @return the set of item names that match the filter
+     */
     public static @NotNull List<String> itemNames(@NotNull String filter) {
         var itemNameFilter = filter.toUpperCase(Locale.ROOT);
         return BoxProvider.get()
@@ -21,6 +30,12 @@ public final class TabCompleter {
                 .collect(Collectors.toList());
     }
 
+    /**
+     * Gets the set of player names that match the filter.
+     *
+     * @param filter the player name filter
+     * @return the set of player names that match the filter
+     */
     public static @NotNull List<String> players(@NotNull String filter) {
         var playerNameFilter = filter.toLowerCase(Locale.ROOT);
 
@@ -32,6 +47,13 @@ public final class TabCompleter {
                 .collect(Collectors.toList());
     }
 
+    /**
+     * Gets the set of player names that match the filter.
+     *
+     * @param filter         the player name filter
+     * @param permissionNode the permission node to check
+     * @return the set of player names that match the filter
+     */
     public static @NotNull List<String> players(@NotNull String filter, @NotNull String permissionNode) {
         var playerNameFilter = filter.toLowerCase(Locale.ROOT);
