+++ features/command/src/main/java/net/okocraft/box/feature/command/boxadmin/StockModifyCommands.java
@@ -8,8 +8,8 @@ import net.okocraft.box.api.message.GeneralMessage;
 import net.okocraft.box.api.model.item.BoxItem;
 import net.okocraft.box.api.model.stock.StockHolder;
 import net.okocraft.box.feature.command.message.BoxAdminMessage;
-import net.okocraft.box.feature.command.util.TabCompleter;
-import net.okocraft.box.feature.command.util.UserStockHolderOperator;
+import net.okocraft.box.api.util.TabCompleter;
+import net.okocraft.box.api.util.UserStockHolderOperator;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
