+++ features/command/src/main/java/net/okocraft/box/feature/command/boxadmin/stock/StockInfoCommand.java
@@ -5,8 +5,8 @@ import net.okocraft.box.api.BoxProvider;
 import net.okocraft.box.api.command.AbstractCommand;
 import net.okocraft.box.api.message.GeneralMessage;
 import net.okocraft.box.feature.command.message.BoxAdminMessage;
-import net.okocraft.box.feature.command.util.TabCompleter;
-import net.okocraft.box.feature.command.util.UserStockHolderOperator;
+import net.okocraft.box.api.util.TabCompleter;
+import net.okocraft.box.api.util.UserStockHolderOperator;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
