+++ features/command/src/main/java/net/okocraft/box/feature/command/boxadmin/stock/StockListCommand.java
@@ -6,8 +6,8 @@ import net.okocraft.box.api.command.AbstractCommand;
 import net.okocraft.box.api.message.GeneralMessage;
 import net.okocraft.box.api.model.stock.StockData;
 import net.okocraft.box.feature.command.message.BoxAdminMessage;
-import net.okocraft.box.feature.command.util.TabCompleter;
-import net.okocraft.box.feature.command.util.UserStockHolderOperator;
+import net.okocraft.box.api.util.TabCompleter;
+import net.okocraft.box.api.util.UserStockHolderOperator;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
